# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /admin/categories:
    get:
      description: OWNER, ADMIN, and STAFF can read paginated categories.
      operationId: AdminCategoriesController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            example: 1
            type: number
        - name: pageSize
          required: false
          in: query
          schema:
            maximum: 100
            default: 20
            example: 20
            type: number
        - name: search
          required: false
          in: query
          schema:
            example: foundation
            type: string
        - name: isActive
          required: false
          in: query
          schema:
            example: true
            type: boolean
        - name: parentId
          required: false
          in: query
          schema:
            example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
            type: string
      responses:
        "200":
          description: Paginated categories.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCategoryResponse"
        "401": &a7
          description: Authentication required.
        "403": &a8
          description: Insufficient admin role.
      security: &a9
        - bearer: []
      summary: List admin categories
      tags: &a10
        - Admin Categories
    post:
      description: OWNER and ADMIN can create categories.
      operationId: AdminCategoriesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryDto"
      responses:
        "200":
          description: Category created.
        "401": *a7
        "403": *a8
        "409":
          description: Category code already exists.
      security: *a9
      summary: Create category
      tags: *a10
  /admin/categories/{id}:
    get:
      description: OWNER, ADMIN, and STAFF can read category details.
      operationId: AdminCategoriesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Category details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCategoryResponse"
        "401": *a7
        "403": *a8
        "404":
          description: Category not found.
      security: *a9
      summary: Get admin category details
      tags: *a10
    patch:
      description: OWNER and ADMIN can update mutable category fields.
      operationId: AdminCategoriesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryDto"
      responses:
        "200":
          description: Category updated.
        "401": *a7
        "403": *a8
        "404":
          description: Category not found.
        "409":
          description: Duplicate code or circular parent.
      security: *a9
      summary: Update category
      tags: *a10
    delete:
      description: OWNER and ADMIN can soft-deactivate categories.
      operationId: AdminCategoriesController_deactivate
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Category deactivated.
        "401": *a7
        "403": *a8
        "404":
          description: Category not found.
      security: *a9
      summary: Deactivate category
      tags: *a10
...
  /orders:
    post:
      description: Creates or reuses a customer, creates a default delivery address,
        creates order item snapshots from a selected recommendation result, and
        creates initial status history.
      operationId: OrdersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderDto"
      responses:
        "201":
          description: Order created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderCreateResponse"
        "400":
          description: Invalid recommendation result, inactive selected
            pack/product/reference, or out-of-stock selected reference.
        "404":
          description: Recommendation result not found.
      summary: Create a Cash on Delivery order
      tags: &a39
        - Orders
  /orders/{id}:
    get:
      description: Returns a limited public order summary. It intentionally excludes
        phone, address, notes, items, and status history.
      operationId: OrdersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Order ID.
          schema:
            example: 00000000-0000-4000-8000-000000000001
            type: string
      responses:
        "200":
          description: Safe public order summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicOrderResponse"
        "404":
          description: Order not found.
      summary: Get safe public order summary
      tags: *a39
  /admin/orders:
    get:
      description: OWNER, ADMIN, and STAFF can search, filter, sort, and paginate orders.
      operationId: AdminOrdersController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            example: 1
            type: number
        - name: pageSize
          required: false
          in: query
          schema:
            maximum: 100
            default: 20
            example: 20
            type: number
        - name: search
          required: false
          in: query
          schema:
            example: ORD-20260612
            type: string
        - name: orderStatus
          required: false
          in: query
          schema:
            type: string
            enum:
              - PENDING_CONFIRMATION
              - CONFIRMED
              - PREPARING
              - SHIPPED
              - DELIVERED
              - CANCELED
              - RETURNED
        - name: paymentStatus
          required: false
          in: query
          schema:
            type: string
            enum:
              - UNPAID
              - PAID
              - REFUNDED
        - name: paymentMethod
          required: false
          in: query
          schema:
            type: string
            enum:
              - CASH_ON_DELIVERY
        - name: selectedPackId
          required: false
          in: query
          schema:
            example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
            type: string
        - name: sourceChannel
          required: false
          in: query
          schema:
            type: string
            enum:
              - INSTAGRAM
              - WHATSAPP
              - TIKTOK
              - FACEBOOK
              - DIRECT
              - OTHER
        - name: createdFrom
          required: false
          in: query
          schema:
            example: 2026-06-01T00:00:00.000Z
            type: string
        - name: createdTo
          required: false
          in: query
          schema:
            example: 2026-06-30T23:59:59.999Z
            type: string
        - name: minTotal
          required: false
          in: query
          schema:
            example: 100
            type: number
        - name: maxTotal
          required: false
          in: query
          schema:
            example: 500
            type: number
        - name: sortBy
          required: false
          in: query
          schema:
            type: string
            enum:
              - createdAt
              - updatedAt
              - totalAmount
              - orderNumber
        - name: sortOrder
          required: false
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
      responses:
        "200":
          description: Paginated orders.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderDetailsResponse"
        "401": &a40
          description: Authentication required.
        "403": &a41
          description: Insufficient admin role.
      security: &a42
        - bearer: []
      summary: List admin orders
      tags: &a43
        - Admin Orders
  /admin/orders/{id}:
    get:
      description: OWNER, ADMIN, and STAFF can view full order details,
        customer/address data, selected pack, items, recommendation summary, and
        status history.
      operationId: AdminOrdersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Complete order details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderDetailsResponse"
        "401": *a40
        "403": *a41
        "404":
          description: Order not found.
      security: *a42
      summary: Get admin order details
      tags: *a43
  /admin/orders/{id}/status:
    patch:
      description: OWNER and ADMIN can move an order through the enforced workflow.
        COD orders are marked PAID on DELIVERED and REFUNDED on RETURNED after
        payment.
      operationId: AdminOrdersController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusDto"
      responses:
        "200":
          description: Order status updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderDetailsResponse"
        "401": *a40
        "403": *a41
        "404":
          description: Order not found.
      security: *a42
      summary: Update order status
      tags: *a43
info:
  title: Beauty Pack Recommendation API
  description: Backend API for beauty-pack personalization, recommendations, COD
    orders, and administration
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Paste the JWT access token returned by POST /auth/login.
  schemas:
...
    AdminSummaryResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        fullName:
          type: string
          example: Admin User
        email:
          type: string
          example: admin@example.com
        role:
          type: string
          enum: &a44
            - OWNER
            - ADMIN
            - STAFF
          example: ADMIN
      required:
        - id
        - fullName
        - email
        - role
...
    AdminCategoryResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        code:
          type: string
          example: FOUNDATION
        name:
          type: string
          example: Foundation
        isActive:
          type: boolean
          example: true
      required:
        - id
        - code
        - name
        - isActive
    CreateCategoryDto:
      type: object
      properties:
        code:
          type: string
          example: FOUNDATION
        name:
          type: string
          example: Foundation
        description:
          type: string
          example: Face foundation products
        parentId:
          type: string
          example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
        imageUrl:
          type: string
          example: https://example.com/foundation.jpg
        sortOrder:
          type: number
          example: 1
          default: 0
        isActive:
          type: boolean
          example: true
          default: true
      required:
        - code
        - name
    UpdateCategoryDto:
      type: object
      properties:
        code:
          type: string
          example: FOUNDATION
        name:
          type: string
          example: Foundation
        description:
          type: string
          example: Face foundation products
        parentId:
          type: string
          example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
        imageUrl:
          type: string
          example: https://example.com/foundation.jpg
        sortOrder:
          type: number
          example: 1
          default: 0
        isActive:
          type: boolean
          example: true
          default: true
...
    CreateOrderDto:
      type: object
      properties:
        recommendationResultId:
          type: string
          description: Recommendation result selected by the customer.
          example: 00000000-0000-4000-8000-000000000001
        fullName:
          type: string
          description: Customer display name for order fulfillment.
          example: Demo Customer
        phone:
          type: string
          description: Customer phone number used to create or reuse a customer.
          example: "0600000000"
        whatsappPhone:
          type: string
          description: Optional WhatsApp phone number for future customer contact.
          example: "0600000000"
        city:
          type: string
          description: Delivery city.
          example: Casablanca
        addressLine:
          type: string
          description: Delivery street/address line.
          example: Maarif
        extraInfo:
          type: string
          description: Additional delivery directions.
          example: Near the pharmacy
        notes:
          type: string
          description: Optional order note for admin fulfillment.
          example: Call before delivery
      required:
        - recommendationResultId
        - fullName
        - phone
        - city
        - addressLine
    OrderItemResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        productNameSnapshot:
          type: string
          example: Foundation X
        referenceNameSnapshot:
          type: string
          example: Medium Warm
        unitPriceSnapshot:
          type: number
          example: 120
        quantity:
          type: number
          example: 1
        totalPrice:
          type: number
          example: 120
      required:
        - id
        - productNameSnapshot
        - referenceNameSnapshot
        - unitPriceSnapshot
        - quantity
        - totalPrice
    OrderCreateResponse:
      type: object
      properties:
        orderId:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        orderNumber:
          type: string
          example: ORD-20260612-0001
        orderStatus:
          type: string
          enum: &a51
            - PENDING_CONFIRMATION
            - CONFIRMED
            - PREPARING
            - SHIPPED
            - DELIVERED
            - CANCELED
            - RETURNED
          example: PENDING_CONFIRMATION
        paymentStatus:
          type: string
          enum: &a52
            - UNPAID
            - PAID
            - REFUNDED
          example: UNPAID
        totalAmount:
          type: number
          example: 299
        currency:
          type: string
          example: MAD
        packName:
          type: string
          example: Natural Glow Pack
        createdAt:
          type: string
          example: 2026-06-12T10:00:00.000Z
        updatedAt:
          type: string
          example: 2026-06-12T10:00:00.000Z
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponse"
      required:
        - orderId
        - orderNumber
        - orderStatus
        - paymentStatus
        - totalAmount
        - currency
        - packName
        - createdAt
        - updatedAt
        - items
    PublicOrderResponse:
      type: object
      properties:
        orderId:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        orderNumber:
          type: string
          example: ORD-20260612-0001
        orderStatus:
          type: string
          enum: *a51
          example: PENDING_CONFIRMATION
        paymentStatus:
          type: string
          enum: *a52
          example: UNPAID
        totalAmount:
          type: number
          example: 299
        currency:
          type: string
          example: MAD
        packName:
          type: string
          example: Natural Glow Pack
        createdAt:
          type: string
          example: 2026-06-12T10:00:00.000Z
        updatedAt:
          type: string
          example: 2026-06-12T10:00:00.000Z
      required:
        - orderId
        - orderNumber
        - orderStatus
        - paymentStatus
        - totalAmount
        - currency
        - packName
        - createdAt
        - updatedAt
    AdminOrderCustomerResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        fullName:
          type: string
          example: Demo Customer
        phone:
          type: string
          example: "0600000000"
        whatsappPhone:
          type: object
          example: "0600000000"
          nullable: true
      required:
        - id
        - fullName
        - phone
    AdminOrderAddressResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        city:
          type: string
          example: Casablanca
        addressLine:
          type: string
          example: Maarif
        extraInfo:
          type: object
          example: Near the pharmacy
          nullable: true
      required:
        - id
        - city
        - addressLine
    AdminOrderStatusHistoryResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        oldStatus:
          type: string
          enum:
            - PENDING_CONFIRMATION
            - CONFIRMED
            - PREPARING
            - SHIPPED
            - DELIVERED
            - CANCELED
            - RETURNED
          nullable: true
        newStatus:
          type: string
          enum:
            - PENDING_CONFIRMATION
            - CONFIRMED
            - PREPARING
            - SHIPPED
            - DELIVERED
            - CANCELED
            - RETURNED
          example: CONFIRMED
        comment:
          type: object
          example: Customer confirmed by phone
          nullable: true
        changedByAdmin:
          nullable: true
          type: object
          allOf:
            - $ref: "#/components/schemas/AdminSummaryResponse"
        createdAt:
          type: string
          example: 2026-06-12T10:00:00.000Z
      required:
        - id
        - newStatus
        - createdAt
    AdminOrderDetailsResponse:
      type: object
      properties:
        orderId:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        orderNumber:
          type: string
          example: ORD-20260612-0001
        orderStatus:
          type: string
          enum: *a51
          example: PENDING_CONFIRMATION
        paymentStatus:
          type: string
          enum: *a52
          example: UNPAID
        totalAmount:
          type: number
          example: 299
        currency:
          type: string
          example: MAD
        packName:
          type: string
          example: Natural Glow Pack
        createdAt:
          type: string
          example: 2026-06-12T10:00:00.000Z
        updatedAt:
          type: string
          example: 2026-06-12T10:00:00.000Z
        paymentMethod:
          type: string
          enum:
            - CASH_ON_DELIVERY
          example: CASH_ON_DELIVERY
        customer:
          $ref: "#/components/schemas/AdminOrderCustomerResponse"
        customerAddress:
          $ref: "#/components/schemas/AdminOrderAddressResponse"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponse"
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/AdminOrderStatusHistoryResponse"
      required:
        - orderId
        - orderNumber
        - orderStatus
        - paymentStatus
        - totalAmount
        - currency
        - packName
        - createdAt
        - updatedAt
        - paymentMethod
        - customer
        - customerAddress
        - items
        - statusHistory
    UpdateOrderStatusDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - PENDING_CONFIRMATION
            - CONFIRMED
            - PREPARING
            - SHIPPED
            - DELIVERED
            - CANCELED
            - RETURNED
          example: CONFIRMED
        comment:
          type: string
          example: Customer confirmed by phone
      required:
        - status
...